# try using LCP data to speed up diff generation