Unfortunately, this was an unsuccessful experiment. Here are the timings without using the LCP data:
```
2.181186 seconds (79 allocations: 1.542 GiB, 9.02% gc time)
0.379120 seconds (104.20 k allocations: 4.885 MiB)
0.311887 seconds (13.30 k allocations: 207.953 KiB)
0.311125 seconds (13.30 k allocations: 207.953 KiB)
```
Here are the timings using the LCP data:
```
6.357530 seconds (75 allocations: 1.691 GiB, 4.06% gc time)
0.419969 seconds (104.20 k allocations: 4.885 MiB)
0.324098 seconds (13.30 k allocations: 207.953 KiB)
0.324188 seconds (13.30 k allocations: 207.953 KiB)
```
So the index generation is much slower, as expected, but the diff also got a bit slower. I tried this with the LCP data interleaved with the suffix array data (as it is now) and with it strictly after the suffix array data and neither was notably better.