@StefanKarpinski 2020-02-28
Worse, this combined with the sk/zrle branch reduces some of the improvements gained by ZRLE-encoding the tarballs before computing the diff. Here's using ZRLE without the LCP array:
```
1.606466 seconds (1.97 M allocations: 640.254 MiB, 9.57% gc time)
0.206462 seconds (165.16 k allocations: 7.519 MiB)
0.100620 seconds (12.35 k allocations: 193.047 KiB)
0.098590 seconds (12.35 k allocations: 193.047 KiB)
```
Here's using ZRLE with the LCP array:
```
1.957610 seconds (1.99 M allocations: 694.268 MiB, 8.93% gc time)
0.239502 seconds (185.67 k allocations: 8.396 MiB)
0.114340 seconds (12.35 k allocations: 193.047 KiB)
0.110325 seconds (12.35 k allocations: 193.047 KiB)
```
So again, it gets a bit slower and of course the index generation is significantly slower.