@StefanKarpinski 2020-02-29
Closing since it didn't result in a speedup.